+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Changelog
 =========
 
+## v1.1.6(JUN 12, 2019)
+ * Added `SendBirdSyncManager.Options` which is used for message count limit of `MessageCollection` and message resend policy.   
+ * Added `handleSendMessageResponse()` in `MessageCollection` to support resend failed message.  
+
 ## v1.1.5(MAY 23, 2019)
  * TypeScript interface fix
  * Improved stability